Organized careers page components into folders
