+++ src/components/careers/desktop/portfolio_section/ProfileSummary.tsx
@@ -6,11 +6,11 @@ import {
   makeStyles,
   Modal
 } from '@material-ui/core'
-import { theme } from "../../../theme";
-import {Member} from "../../../constants/members";
-import {ChipSmall} from "./ChipSmall";
+import { theme } from "../../../../theme";
+import {Member} from "../../../../constants/members";
+import {ChipSmall} from "../ChipSmall";
 import {PortfolioButton} from "./PortfolioButton";
-import {ProfileFull} from "./profilePopUp/ProfileFull";
+import {ProfileFull} from "../profilePopUp/ProfileFull";
 
 const BUTTON_TEXT = 'PORTFOLIO';
