+++ src/components/careers/desktop/portfolio_section/ProfileWheel.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import {Box, Grid, makeStyles, styled} from '@material-ui/core'
-import {Member} from "../../../constants/members";
+import {Member} from "../../../../constants/members";
 import {ProfileSummary} from "./ProfileSummary";
