+++ src/components/careers/desktop/testimonial_section/GreenDotsRectangle.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import {styled, Grid} from '@material-ui/core'
-import {theme} from "../../../theme";
+import {theme} from "../../../../theme";
 
 const StyledGrid = styled(Grid)({
   height: 'max-content',
