+++ src/components/careers/desktop/testimonial_section/TestimonialCarousel.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import {Box, makeStyles} from '@material-ui/core';
-import {Testimonial, Testimonials} from "../../../constants/testimonials";
+import {Testimonial, Testimonials} from "../../../../constants/testimonials";
 import {TestomonialItem} from "./TestomonialItem";
 import Carousel from 'react-material-ui-carousel';
